Reintroduce user creation in create_dev_data script
